+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.2.13 (2023-12-08)
+* Added allowed headers to CORS Middleware
+* Added JSON error response in ErrorHandler Middleware
+
 ## v0.2.12 (2023-12-08)
 * Added CORS Middleware
 * Added priority ordering to incoming Middleware